requirements.txt
Django==4.1.3
whitenoise==6.2.0
django-filter==23.2
django-model-utils==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==9.5.0
django-cors-headers==4.2.0
